+++ src/gtirb_pprinter/Registration.cpp
@@ -54,6 +54,7 @@ void registerAuxDataTypes() {
   gtirb::AuxDataContainer::registerAuxDataType<ElfDynamicFini>();
   gtirb::AuxDataContainer::registerAuxDataType<ElfStackExec>();
   gtirb::AuxDataContainer::registerAuxDataType<ElfStackSize>();
+  gtirb::AuxDataContainer::registerAuxDataType<ElfSoname>();
 }
 
 void registerPrettyPrinters() {